chore(01): use regex instead of splits
I like this approach a little more. Turns out using Regex for group
matches is much faster than splutting the string into an array. Like 40%
faster.
I don't like destructuring the groups into an array, then having to sort
the contents into separate lists. It's just as bad as the original. So
really our speed up is solely that first split of the array. Probably
some small gains in the forEach to lists, but I doubt much.

Overall faster, but still not pleasant. I would love to find an
algorithm or effect way to sort the lists in place or sort in order on
push.