chore(01): pt2 loop with mutating addition
Continuing pt2 with the regex lookup from pt1alt, because I am pleased
with the performance. Scoring simliarity is based on how many times the
a left number appears in the right column, then summed.
Doing this in a for loop with a filter means we'll iterate over the
right list, left times. which is gonna be expensive.