chore(01): pt2 reducer with dynamic regex match
I thought this approach would be faster. We skip the filtering instead
using regex to match all of the occurrances of left. Turns out this is
incredibly slow. I think it might be the creation of the regex each
time. There might be a way to optimize that, however I think there will
be limited returns. A more simple regex, with less matchers, can reduce
the cost substantially. Like removing \\d+ and + from \\W+ improves
speed by 10x. This is a good lesson to make Regexs as simple as
possible, with as few constraints as needed to be accurate.