chore(01): pt2 count occurrances while creating lists
I was disappointed with the performance of the scoring functions for
pt2. While working on the actual data from the day;

Filter method = 9.162ms
Regex method with reducer = 22.279875ms
Regex method with reducer and bad Regex = ~220ms

Another option I thought about was to use a map to store the key and of
the number and the number of times it's been seen. This way we only
iterate the occurrances list once, and everything is done before each
loop iteration so the loop execution is short.

Maps may still be a slow data structure (I don't know how memory is
allocated, I assume like a Set, which is all over the place). This
approach is much faster than the filter method on the large dataset

Map: 0.545417ms

I am much happier with that time.