chore(02): iterate over array of levels with direct compare
Soluton for part 1 of day 2, achieved by iterating over each level and
performing the comparison.

The work is straight forward and the code is performant. Coming in
around 0.02ms.

This is infact my second attempt. The first attempt hard coded the idea
of 5 levels per report, so when run against actual data it returned 29
extra false positives.

I think the code is neat, but a little verbose. Having the loops defined
for increase and decreasing is tedious. And leaves room for error if you
need to go in an chage something.

Even as simple as changing the iterator symbols i and j. They're too
similar which coud cause a failure using the wrong variable. The easy
solution is to use words or dissimilar characters.