chore(02): pt2 working
I struggled so hard with this. Even with the bruteforce nature of the
approach, I struggled. Turns out there was also some trash data in the
file that needed to be trimmed so the last while I was working on fixes,
I was missing that step. A good reminder to sanitize your data before
doing operation.

Another good reminder is lots of test cases. Lots and lots. I cheated
here and used the subreddit to find other people test cases, and even
then I still wasn't successful - I think again this was due to the bad
data.