chore(04): working implementation
It is by no means quick, however easy enough to brute force looking at
each X and generating a list of coordinates to check.

For some reason, the guard clauses on the directions generator were
causing a miscount of 17 instead of 18 on the sample data. This should
be investigated.

I can't really think of ways to improve this, but I'm pretty sure there
is an way to more efficiently map the directions to check, perhaps
returning the value at the same time - why store the coord  multple
times when we can do the check in the same pass.