chore(04): pt1 cleanup pass
The number of times we pass over the data can be reduced substantially.

Instead of generating coordinates, and then going through each set to
find the letter, then going through each set to check the word - we can
check the word in the first pass. It still requires looking up each
letter, but now its done in 1 loop.

This also led to the realization we don't need to worry about the
boundaries of the 2dArray through Max and guard clauses. Yay.