chore(04): pt1 cleanup pass 3
It's commonplace to dislike reducers due to their syntax and the
complexity understanding them involves. In this case, it is much more
cryptic than the for loop with accumulator, but internal is small enough
that it should be reasonable to keep in mind. But still, a little more
complexity might not be worth it.

For _some_ reason, this performs faster on sample and actual data,
bringing the timing of sample runs down to 1.6ms which is about what it
was in the working implementation. Actual data operation is down to
10ms. Which from working implementation was 25ms. These gains weren't
realized entirely in this refactor, but moving to the reducer shaved off
5ms or so.