chore(04): pt1 cleanup pass 4
The collection of directional functions to create the coordinates was
bugging me (after this pass it might be worse).

Placing them all in a single function with the matrix computed up up
front, we reduce the number of calls where we're passing the
input2dArray ref and coordinates around which is great.

This is infact slower on both sample and actual data thought. I wonder
if it's because the operations can't be optimized with the V8?