chore(04): pt2 cleanup pass
I got carried away between changes.

First order of business is to replace ['M', 'S'].includes check with
direct === 'M' === 'S' checks which are faster for 2 entries.

Next was to question why we are filting at all instead of just appending
any value (incuding undefined) to the string of letters. The switch is
pattern matching, which makes things simple, so we can relax our guards
and preparing the strings for it.

This means we can also remove the results and count checks, directly
using the switch, and within the switch body we can use count instead of
accumulating words, and checking length later.

Using a reducer to accumulate the letters is short and simple enough so
I think it's ok here (especially compared to the pt 1 reducers).

Overall this pass brough the compute time from 0.4ms sample data and 8ms
for actual data, to 0.29ms sample, 2.6ms actual data. 3x increase in
perormance is a big in here. Even if the sample data is only 25% faster.