chore(04): both cleanup pass
Last thing that was bugging me was performing the input to 2d array,
then iterating the matrix to find coordinates of target letters. This
can be reduced to being done inline with the creation of the 2d array,
where the lookups are only done against the length of each row.

This doesn't really change performance on either dataset, but is less
code, so maybe that is a win for maintainability.