chore(05): pt 1 working implementation
My approach, like Day 04, was to start with a plan.

The plan is simple, parse the data into friendly data structures, like
an object with keys of pages, and values of the pages that must come
after.

The planned approach works, but is a lot of loops. For one of the
examples, the update is 3 pages, with 6 conditions to check. That's a
total of 18 operations. The actual data has lots, making checking each
number very inefficient.

Something cool that I've never used before is the `label:` with for
loops. This allows us to break and continue within nested loops. Because
I've just learned about this, I want to use it everywhere, but am
worried about the syntax being too unfamiliar. I'm shocked I've never
encountered it before.

On to part 2. I hope we'll realize an optimization for part 1 through
it.