chore(05): pt2 working implementation
Moving elements in an array is tedious. I don't like splice. Hopefully
we'll get a Array.prototype.move in the future. Preferrably a
not-mutable version (yes we can add our own, but I don't like modifying
prototypes).

This approach is alright, I wonder if there is something better. Like,
do we have to sort all the numbers? or do we just have to ensure the
middle number is valid?