chore(06): pt1 working implementation
I went down the path of removing the while loop, and depending solely on
recursion. This worked for the sample data, however if crashes on the
actual data. So I brought back the while loop and moved the ledger to
record the good cursor before the nextCursor (which was the cause of the
off by one error)