chore(07): pt1 dead end?
So I thought I was off to a good start with this approach. Generating
all the possible operators for each test value and then performing and
checking. This seemed like a good idea, but runs rather slow. It solves
the sample data just fine, but the actual data for part 1 fails.