chore(07): pt2 working implementation
Using the array of operations works really well for extending with new
operations. I think if the original approach would have exploded
horribly. Yay.

I think we can reduce the duplication between the methods, as long as we
can instruct the solver function to add the third operation to the
array. But otherwise this is done.