chore(07): cleanup recursion syntax
This looks more confusing but I do like how passing `...numbers` writes.
But because it's code we wrote, and control, it's fine. I'm glad this
feels like a smell, because we should never pass more arguments than
needed to a function we don't control, incase it does things we don't
realize, or changes.