chore(07): refactor everything
I wasn't a fan of the reducer being initialized with the zero at the
end of the lines. A solution for that is to put the 0 at the start of
the array for the reducer. In doing this I moved the Number map to
inside the array split, still iterating over the same count of items as
if being on a separate pass so it's no better or worse

I expended perform to applyOperations to be more descriptive of what the
function is doing. I don't mind because...

I shorted the name of the variables, hopefully the context is not lost
when reading the whole method. By that, I mean line by line might be
harder to read.

Something exciting happened this pass.

Naming head and next variables in the applyOperations signature instead
of splitting them out after and referencing a few times, actually
performs around 25% faster. This is attributed to not having to mutate
the numbers array with slice.
using slice(2)
part 1: sample data    0.732042ms
part 1: actual data   17.272708ms
part 2: sample data    0.160250ms
part 2: actual data  517.592450ms

head/next without slice(2)
part 1: sample data    0.577208ms
part 1: actual data   10.504792ms
part 2: sample data    0.161667ms
part 2: actual data  351.668080ms